Bahkses!
Going on with my idea of integrating previous projects into this one, I
created a remastered version of my 3D collisions project, which amounts
for size, as opposed to the previous one, which could only do 50x50x50
collisions. I also added physics to the camera, accounting for
location, velocity, and acceleration, as to allow for better movement.

+added collisions
+added physics